Start the app in SwiftUI
